+++ server/tests/api/videos/video-abuse.ts
@@ -2,6 +2,7 @@
 
 import * as chai from 'chai'
 import 'mocha'
+import { VideoAbuse } from '../../../../shared/models/videos'
 import {
   flushAndRunMultipleServers,
   flushTests,
@@ -83,11 +84,11 @@ describe('Test video abuses', function () {
     expect(res1.body.data).to.be.an('array')
     expect(res1.body.data.length).to.equal(1)
 
-    const abuse = res1.body.data[0]
+    const abuse: VideoAbuse = res1.body.data[0]
     expect(abuse.reason).to.equal('my super bad reason')
-    expect(abuse.reporterUsername).to.equal('root')
-    expect(abuse.reporterServerHost).to.equal('localhost:9001')
-    expect(abuse.videoId).to.equal(servers[0].video.id)
+    expect(abuse.reporterAccount.name).to.equal('root')
+    expect(abuse.reporterAccount.host).to.equal('localhost:9001')
+    expect(abuse.video.id).to.equal(servers[0].video.id)
 
     const res2 = await getVideoAbusesList(servers[1].url, servers[1].accessToken)
     expect(res2.body.total).to.equal(0)
@@ -111,27 +112,27 @@ describe('Test video abuses', function () {
     expect(res1.body.data).to.be.an('array')
     expect(res1.body.data.length).to.equal(2)
 
-    const abuse1 = res1.body.data[0]
+    const abuse1: VideoAbuse = res1.body.data[0]
     expect(abuse1.reason).to.equal('my super bad reason')
-    expect(abuse1.reporterUsername).to.equal('root')
-    expect(abuse1.reporterServerHost).to.equal('localhost:9001')
-    expect(abuse1.videoId).to.equal(servers[0].video.id)
+    expect(abuse1.reporterAccount.name).to.equal('root')
+    expect(abuse1.reporterAccount.host).to.equal('localhost:9001')
+    expect(abuse1.video.id).to.equal(servers[0].video.id)
 
-    const abuse2 = res1.body.data[1]
+    const abuse2: VideoAbuse = res1.body.data[1]
     expect(abuse2.reason).to.equal('my super bad reason 2')
-    expect(abuse2.reporterUsername).to.equal('root')
-    expect(abuse2.reporterServerHost).to.equal('localhost:9001')
-    expect(abuse2.videoId).to.equal(servers[1].video.id)
+    expect(abuse2.reporterAccount.name).to.equal('root')
+    expect(abuse2.reporterAccount.host).to.equal('localhost:9001')
+    expect(abuse2.video.id).to.equal(servers[1].video.id)
 
     const res2 = await getVideoAbusesList(servers[1].url, servers[1].accessToken)
     expect(res2.body.total).to.equal(1)
     expect(res2.body.data).to.be.an('array')
     expect(res2.body.data.length).to.equal(1)
 
-    const abuse3 = res2.body.data[0]
+    const abuse3: VideoAbuse = res2.body.data[0]
     expect(abuse3.reason).to.equal('my super bad reason 2')
-    expect(abuse3.reporterUsername).to.equal('root')
-    expect(abuse3.reporterServerHost).to.equal('localhost:9001')
+    expect(abuse3.reporterAccount.name).to.equal('root')
+    expect(abuse3.reporterAccount.host).to.equal('localhost:9001')
   })
 
   after(async function () {